+++ testproject/testproject/models.py
@@ -3,7 +3,8 @@ from django.db.models.signals import post_save, post_delete
 
 from mptt.models import TreeManager
 
-from groups_manager.models import Group, GroupType, group_save, group_delete
+from groups_manager.models import Group, Member, GroupType, \
+    group_save, group_delete, member_save, member_delete
 
 '''
 Legion main example
@@ -112,6 +113,49 @@ class WorkGroup(Group):
 
 post_save.connect(group_save, sender=WorkGroup)
 post_delete.connect(group_delete, sender=WorkGroup)
+
+
+'''
+Proxy Group and Member
+'''
+
+
+class OrganizationMember(Member):
+
+    class Meta:
+        proxy = True
+
+post_save.connect(member_save, sender=OrganizationMember)
+post_delete.connect(member_delete, sender=OrganizationMember)
+
+
+class Organization(Group):
+
+    class Meta:
+        proxy = True
+
+    class GroupsManagerMeta:
+        member_model = 'testproject.OrganizationMember'
+
+post_save.connect(group_save, sender=Organization)
+post_delete.connect(group_delete, sender=Organization)
+
+
+class OrganizationMemberSubclass(Member):
+    phone_number = models.CharField(max_length=20)
+
+post_save.connect(member_save, sender=OrganizationMemberSubclass)
+post_delete.connect(member_delete, sender=OrganizationMemberSubclass)
+
+
+class OrganizationSubclass(Group):
+    address = models.CharField(max_length=200)
+
+    class GroupsManagerMeta:
+        member_model = 'testproject.OrganizationMemberSubclass'
+
+post_save.connect(group_save, sender=OrganizationSubclass)
+post_delete.connect(group_delete, sender=OrganizationSubclass)
 
 
 class Pipeline(models.Model):