+++ docs/source/changelog.rst
@@ -1,4 +1,8 @@
 Changelog
 =========
 
-- 14-10-29: Initial version (0.1.0)
+- 15-05-05 (0.2.0):
+    - Changed retrieval of permission's name: 'view', 'change' and 'delete' will be translated to '<name>_<model_name>', the others are left untouched (see :ref:`permission name policy <permission-name-policy>`)
+    - Added GroupsManagerMeta class to Group that allows to specify the member model to use for members list (see :ref:`custom Member model <custom-member-model>`)
+
+- 14-10-29 (0.1.0): Initial version